66 update join ikl

+++ ft_irc/src/core/Channel.cpp
@@ -125,7 +125,7 @@ bool Channel::hasUserLimit() const
     return _hasUserLimit;
 }
 
-int Channel::getUserLimit() const
+size_t Channel::getUserLimit() const
 {
     return _userLimit;
 }
